Distributor: replace ring.DoBatch with ring.DoBatchWithClientError
